Define methods that create CF template
Several methods of the IAMTemplateGenerator class were defined
that generate a CloudFormation template file. This template
file defines a new IAM user with the permissions needed to deploy
a static website to AWS using S3 and CloudFront.

A YAML file containing the expected contents of this template was
added to the tests directory. It is used by one of the tests to
compare against the template generated by the IAMTemplateGenerator
class.