Define CloudFormation template resources
Several classes were created that add AWS resources to the
CloudFormation template. These classes define an S3 bucket that
is meant to store the static files of the website, a CloudFront
distribution that serves the files in the S3 bucket, DNS record
sets that point the site's domain name to the CloudFront
distribution, and an SSL certificate provisioned using AWS
Certificate Manager.