Defines class that creates AWS CF stack
A new class was defined that will eventually create a new
CloudFormation stack using the template produced by the
definitions module. Right now, this new class only retrieves the
hosted zone ID of the hosted zone for the domain name passed as
a command-line argument to the main.py script. The methods that
create a new CloudFormation stack still have to be written.