Define method to create CloudFormation stack
A new parent class of the CloudFrontDistributionStackCreator class
was defined. This parent class has several utility methods that
can be used to create a new CloudFormation stack and check on the
status of the stack creation. Also, a new method of the
CloudFrontDistributionStackCreator was defined that creates a new
stack. This method is called any time the main.py script is run
with the 'deploy' command-line argument.