Refactor project to use configuration file
The entire project was refactored to use a configuration file
instead of command-line arguments. The only command-line arguments
that are still accepted are the 'action' argument and a new
optional argument called 'config' that accepts the name of the
configuration file to use.

The number of command-line arguments was getting cumbersome, so
it was decided that a configuration file would be more
convenient. An example configuration file has been included in
this repo that describes the available settings.