Remove Certificate resource from CF template
The CloudFormation template that creates the resources needed to
host the static site no longer provisions an SSL certificate
using AWS Certificate Manager. Using CloudFormation to provision
certificates and perform DNS validation of these certificates
often results in a race condition where the resources that
depend on the certificate can't be created while the certificate
is in the pending validation state. This resulted in problems
with the CF stack being created.

Therefore, the Certificate resource has been removed. A new method
is going to be written that will request a new certificate and
validate it before the CF stack is created. This synchronous
approach is more failproof.