Allow settings file to reside in other directories
Previously, the Arguments class in the validators.py module
required the settings file to be located in the project's root
directory. This was because the importlib.import_module function
requires Python modules to be located in a directory that is
accessible from the root directory. This has been changed and now
the settings file may be located anywhere on the filesystem.