Ensure correct error pages are used
The CloudFormation template specifies the files in the S3
bucket to use as the default 404 and 500 responses. Two hardcoded
values of '404.html' and '500.html' were previously being used
without regard for whether or not these files actually existed.

This has been changed. Now, the values of the _404_FILE and
_500_FILE settings are used to specify the paths to the error
files. This ensures that these files actually exist in the S3
bucket.